Wiremock

+++ src/main/java/guru/springframework/brewery/web/model/BeerPagedList.java
@@ -32,13 +32,14 @@ public class BeerPagedList extends PageImpl<BeerDto> {
     public BeerPagedList(@JsonProperty("content") List<BeerDto> content,
                          @JsonProperty("number") int number,
                          @JsonProperty("size") int size,
-                         @JsonProperty("totalElements") int totalElements,
+                         @JsonProperty("totalElements") Long totalElements,
                          @JsonProperty("pageable") JsonNode pageable,
                          @JsonProperty("last") boolean last,
                          @JsonProperty("totalPages") int totalPages,
                          @JsonProperty("sort") JsonNode sort,
                          @JsonProperty("first") boolean first,
                          @JsonProperty("numberOfElements") int numberOfElements) {
+
         super(content, PageRequest.of(number, size), totalElements);
     }
